+++ src/main/php/lang/Properties.php
@@ -38,6 +38,12 @@ class Properties implements \Iterator
      */
     public function __construct(array $propertyData = [])
     {
+        foreach ($propertyData as $section => $values) {
+            if (isset($values['password'])) {
+                $propertyData[$section]['password'] = SecureString::create($values['password']);
+            }
+        }
+
         $this->propertyData = $propertyData;
     }